void reverseEntreCuts(ListaPonto* solucao, int cut1, int cut2, ListaPonto* entrada)

+++ ponto.c
@@ -21,7 +21,7 @@ void insereDemanda(Ponto * ponto, double demand){
     ponto->demand = demand;
 };
 
-double retornId(Ponto * p){
+int retornId(Ponto * p){
     return p->id;
 }
